FULL REWRITE: make the themes more programmatic
This provides the basis for the sustainable development of the Modus
themes.  The colours have been consolidated into a consistent palette,
while the colour combinations have been thoroughly reviewed to conform
with the highest accessibility criterion for colour contrast (between
background and foreground values).

Both `modus-operandi' (light) and `modus-vivendi' (dark) now share the
same template.  Only difference are the colour values.  This means
that the mapping of colours to objects is consistent, making the
transition between the two smoother.  The Modus themes can now be
considered part of the same group, which was not really the case in
their initial state.

The development effort will now shift to the coverage of more faces.
Priority will be given to the most common ones and to those provided
by packages I use or have been exposed to.

The Modus themes already are in a workable state.  I consider them
"very good" for my specific needs.  However, I am aware that there
still is room for improvement and potential adjustments.  As such, I
will not be providing a packaged version just yet.