Support for ivy and swiper
Note that the colours for ivy-minibuffer-* and swiper-*-match are not
symmetrical in terms of what their sequential naming would suggest.
This is because these faces are not designed to be mirrors of one
another.  My colour definitions are designed to make them look the
same.  This makes the swiper matches on the buffer and the minibuffer
look the same.