Revise org-block line styles (extends 258b7d4)
The begin and end lines of an org-block no longer have a distinct
background compared to their content.  This is necessary to ensure
that the highlighted line can remain visible at all times.

Note that the hl-line uses a background value that was previously
assigned to the begin/end lines—and this value cannot be much
different, otherwise all colours rendered on top of it might fall
below the accessibility threshold of 7:1 contrast.