Modus vivendi: several palette refinements (read commit message)
The purpose of these changes is four-fold:

1. Ensure consistency of magenta{,-alt,-alt-other} with their
counterparts in Modus operandi.  This practically means that the first
is closer to pink (more red than blue), the second is towards
purple (~equal red+blue), the third is violet (more blue than red).

2. Improve the reds to make them semantically correct.  They were
closer to magenta or yellow.

3. Improve the combination between cyan-alt-other and blue-alt-other
as well as the latter with the magentas.

4. Slightly desaturate and/or adjust {green,yellow}-alt-other.  This
is because they are used as headings, e.g. in org-mode, where their
intensity is not consistent with the rest of the theme.