UPGRADE: colours and faces for "refined" combos
These are styles for specialised faces where both the foreground and the
background need to have the same or similar hue, at the appropriate
contrast ratio of >= 7:1.

They are currently implemented in the git diff faces, making it easier
to denote in-line changes (rather than just marking the entire line as
added/removed/changed).

For Magit to support these faces, you need this:

	(setq magit-diff-refine-hunk t)

Or to mark them all at once:

	(setq magit-diff-refine-hunk 'all)

* * *

This commit greatly expands upon 1554e22.