Modus operandi: palette refinements
The idea with these is to further enhance the differences between the
magentas.  The blues are also tweaked to be more distinct from each
other, but also to improve their contrast vis-à-vis the magentas (the
"purple" variants, to be precise—the different shades do not have their
own name in the interest of simplicity).