Support all-the-icons (but see commit message)
This package defines faces using colour names with dark/light
variations.  Due to the overarching objective of the Modus themes (WCAG
AAA accessibility) it is not possible to support all the colours in that
range.  For example, a dark blue foreground (the icon) on a black
background is not an accessible combination.  Same with many others…

What I have defined here is a good approximation (typos or errors
notwithstanding).