Review `org-warning' face
The intense red background distorts the meaning of this face in contexts
such as the Org agenda view.  Whereas this patch improves things in that
regard.

The reason for going with the original design is that `org-warning' is
also used by `org-export-dispatch' (C-c C-e).[0] That specific interface
should ideally have a dedicated face, in order for us to be able to use
an intense colour combination there, without causing problems to other
Org interfaces.

[0]: https://gitlab.com/protesilaos/modus-themes/issues/2