Numerous refinements to Org faces
An emphasis was placed on the semantic use of colour in the agenda
view (`org-agenda') and on how to best convey the meaning of the various
components of that interface.

All subtle backgrounds from keywords and other constructs were removed,
because they created inconsistencies under certain circumstances.

Other minor refinements were introduced in the interest of harmonising
everything to the new styles.