General usability refinements for ACE-WINDOW
The problem with the previous faces was that the overlay character would
overflow the entire line with its background colour, if that line was
empty.  Now that the background is set to be the same as the main
background, this no longer creates major inconsistencies.

Furthermore, the rest of the window is dimmed entirely, including all
foreground values.  That way it is easier to recognise the overlay
character that `ace-window' presents.

The overlay's foreground value has been changed from the overall green
looks it had before to an appropriate red, with the idea of conveying a
sense of urgency.