Review all cursor faces for robustness
Quoting from the help page of the `cursor' face:

	Basic face for the cursor color under X.  Currently, only the
    ‘:background’ attribute is meaningful; all other attributes are
    ignored.  The cursor foreground color is taken from the background
    color of the underlying text.

	Note: Other faces cannot inherit from the cursor face.

Given this information, I am reverting the `cursor' to use just the main
foreground value.  This means that for Modus Operandi (light theme) the
cursor is pure black.  For Modus Vivendi (dark theme) it is pure white.

For all other faces that control the presentation of the cursor, I am
explicitly defining inverted background and foreground values, instead
of using the unsupported `:inherit cursor'.

The cumulative effect of these changes should hopefully address the
underlying problems present in issues #6 and #7.