Add reserved colours for `whitespace-mode'
These are designed to offer subtle nuances between interfaces that may
also be active, such as `line-number-mode' and `hl-line-mode'.

That granted, it is impossible to create "invisible" characters while
also conforming with the overarching accessibility criterion of the
Modus themes (minimum contrast ratio of 7:1, which is the WCAG AAA
standard).