Subtle tweak to `lazy-highlight' face
This is meant to improve the lazily highlighted matches of
`isearch' (and relevant commands) under certain circumstances.  What
changes in particular, is a more accented (cyan-ish) foreground, so that
the lazy matches appear slightly more dimmed.

All this without prejudice to the overarching accessibility objective of
this project to ensure a minimum contrast ratio of 7:1 (the highest
standard of its kind).