Remove explicit `:extend t' property
This concerns a new property that comes in Emacs 27, which controls
whether a face should extend beyond the column length of the text.

Earlier versions of Emacs 27 had a backward-incompatible behaviour that
practically required themes to be explicit about which faces are to be
extended.

This has now been reviewed upstream,[0] so there is no need for themes
to keep such complex code in place.[1]

[0]: https://git.savannah.gnu.org/cgit/emacs.git/commit/?id=58fb4c3e68a4a42ad491d0fa2c084e5c39942e2b
[1]: https://github.com/alphapapa/ya-solarized.el/issues/1#issuecomment-568295275