Implement new colours/faces for "special calm"
As with all "*-special-*" values, these are meant to be used only in
cases where the main colours would be inappropriate.

The idea is that the "special" set is closer to the grayscale than the
standard accent colours.  As such, they can perform better in
circumstances where the standard white/black combinations would be
semantically incorrect, while an accent colour would be a bad choice as
well.

For example, the "warm" and "cold" special values are used in Magit's
diff buffers to style structural elements, because the use of proper
red/yellow/magenta or blue/cyan/green for those would create confusion
as to what really demands the user's attention (which is the diff
itself).