Review `match' face
This is supposed to make the `occur' buffer (or others inheriting from
it) less intense, while retaining the good visibility of the original
matching terms.

The problem with the previous setting was that the colours were more
pronounced than what their role demanded.  For a buffer with lots of
results, this could make working with the text quite distracting.

Furthermore, the previous colours would make it difficult to
disambiguate the original match from lazily highlighted ones, in the
case of `isearch', `query-replace', etc.