Dedicated colours for paren-match and refinements
The intention behind this addition is to allow matching parentheses to
be clearly distinct from their context.  The previous use of magenta
could potentially make the highlight appear like a code construct or the
match of some search method.

To make the visuals consistent, the highlighting of balanced expressions
is reviewed to approximate that of the matching delimiters.  I have
opted not to use the same colours for this one, as those would be more
intense than their role demands (whereas delimiters can afford to be
intense because they are single characters).