Update Ivy and Swiper current matching line
These changes remove the previous design of drawing a box around the
current line.  That caused subtle inconsistencies whenever the line
would be visually wrapped.

To account for the lack of a box, the background colour of the line is
reviewed in an attempt to offer improved contrast.

Finally, changes have been made to accomodate the transition to Emacs
27, so that the styles for the matching line extend to the end of
window, rather than stopping at the last character.  Without specifying
`:extend t' we easily encounter scenaria where switching between a list
of candidates (by moving up or down) will create a "bounce" effect, as
the line expands and contracts to match the character length.