Fix outline-mode faces' dependency on org-mode
I expect this to solve issue #9.

These faces were originally designed to inherit from the eight heading
styles of org-mode.  However, this created an undesirable dependency,
where enabling `outline-mode' BEFORE loading Org would fail to apply
any styles to the headings.

That is to be expected since in that scenario we were telling Emacs to
inherit from yet-to-be-known faces.

Steps to reproduce:

1. In a shell type `emacs -Q'
2. M-x load-theme RET modus-vivendi RET (or modus-operandi)
3. M-x outline-mode (no styles for headings)
4. Open an org-mode file (or M-x org-mode).
5. Outline mode now also has styles as well.

Thus the solution is to apply styles directly to each group without
inheriting from other groups that may or may not be part of a user's
setup.  Inheritance should only be limited to the custom `modus-*'
faces (other commits may be necessary to make sure that no other bugs
of this sort are encountered).

By the way, the syntax in this commit is more complex because it
accounts for the user-facing customisation options (all of which are
off by default).  See the README for more on them.