Review face inheritance to avoid possible bugs
The rationale for this change was already outlined in commit b02feae.
Basically the point is to be more careful with inheriting styles due
to the implicit dependencies they create.

Inheritance should only be limited to the custom `modus-*' faces that
are part of each theme.  From my perspective these are only known
constant in a user's setup.  One face group should not inherit from
another group because there may be cases where the relevant library is
not loaded.

An exception is made for basic faces such as "error" and "success",
though I will review those as well if the need arises.