Several palette refinements
Most changes are virtually imperceptible.  There are, however, a few
that are quite noticeable: those were necessary to make the colours
more consistent with the rest of the palette and improve their
function in various interface elements.

All refinements were made to improve the disambiguation of visual
elements under several scenaria of interaction.

Lots of possible combinations were put to the test in order to also
account for theoretical outcomes: no actual intersections exist
between such face groups.  Still, it is important to anticipate such
states of affairs.

In practice: colours are easier to tell apart and are supposed to
better convey their meaning.