Further refinements to GNUS group level colours
This fixes a potential issue with commit aaca4a2 where an empty mail
group level 1 could be mistaken for an empty level 4 news group.

Overall, these tweaks should make all items in the "Group" buffer easier
to tell apart, including the cases where mail and news sources are
placed next to each other and are sorted in such a way as to mix things
together.

In my tests the potential of conflating EMPTY group levels only existed
when some mixed sorting was applied (not level-dependent) AND all groups
were turned visible WITHOUT the use of any topics.  With the present
commit I hope to have effectively addressed even that kind of problem.

Note though that the differences in colour were still designed with the
intention of avoiding a pronounced "rainbow effect".  So the contrast
between each colour is, in principle, not too intense (with the obvious
exception of combinations like blue and yellow that are naturally far
apart on the scale).