Add support for the built-in WINDOW-DIVIDER-MODE
The mode is defined in the `frame.el' library (M-x find-library frame).

The default width of the divider is 6 pixels.  The first and last pixels
are meant as narrow outlines that are drawn in a slightly more intense
colour than the remaining 4 inner pixels.

Due to the possibility of adjusting the thickness of the divider, plus
the generous default width, the colours I have chosen are less intense
that those used in the default window border (those are black on white
and vice versa, as per the value of the `default' face).

As such if you wish to have a more subtle window divider that retains
the looks of the generic window border, you can evaluate this (check the
other variables as well):

    (setq window-divider-default-right-width 1)

And then enable the mode:

    (window-divider-mode 1)