Minor refinements to ORG
The `org-headline-done' needs a colour that is consistent with its
semantics.  To apply this face, you need to evaluate:

    (setq org-fontify-done-headline t)

May also need to restart Org:

    M-x org-mode-restart

The `org-tag' must change to accommodate the above, while also remaining
distinct from other constructs in its vicinity.

While the `org-time-grid' is something I missed in commit 0fc2862.