Org Agenda: scalable headers, proportional fonts
Each of the themes has optional support for scalable headers and the use
of proportionately-spaced fonts in those same elements.  Until now, this
was only applied to Org's headlines (and a few other places).  It now is
extended to the agenda views as well.

These are OFF BY DEFAULT: you need to explicitly set the relevant
variables to `t'.  Those are:

    modus-operandi-theme-proportional-fonts
    modus-operandi-theme-scale-headings

    modus-vivendi-theme-proportional-fonts
    modus-vivendi-theme-scale-headings

And this is the default scale:

    modus-operandi-theme-scale-1 1.05
    modus-operandi-theme-scale-2 1.1
    modus-operandi-theme-scale-3 1.15
    modus-operandi-theme-scale-4 1.2

    modus-vivendi-theme-scale-1 1.05
    modus-vivendi-theme-scale-2 1.1
    modus-vivendi-theme-scale-3 1.15
    modus-vivendi-theme-scale-4 1.2

The `org-agenda-date-today' was also tweaked in the interest of
consistency.