Refine `font-lock-warning-face' (see commit msg)
While this is a small change, it actually is not as straightforward.

Normally `font-lock-*' faces are used _directly_ inside a buffer, but
not on the mode line (they can be used _indirectly_ by other faces that
inherit them).  An exception to this norm is `font-lock-warning-face'.

The background values of a buffer and of the {in-,}active mode line vary
considerably.  The latter uses the most pronounced background out of all
those that are meant to present accent colours (red, green, blue, etc.).

To conform with the overarching objective of this project with regard to
the WCAG AAA accessibility standard (contrast ratio >= 7:1), it is
necessary to choose colour combinations that take the background value
as a known constant.

Put differently, it is necessary to have accent values that are intended
for use in the mode line (or rather, to be combined with the mode line
backgrounds).  The subset of colours whose name included the `-active'
suffix covers this particular need.

Couched in those terms, the use of `yellow-active' as a foreground
ensures that this face can appear in the mode line as well be cast on
all background values specified by my themes.

This commit addresses issue 20:
https://gitlab.com/protesilaos/modus-themes/-/issues/20