MAJOR REVIEW of diff views (`diff-mode', `magit')
+ Define an entire new subset of the colour palette.  This one is
designed specifically for diffs in their various states.

+ Define bespoke faces that apply the new colours.  These are internal
to each theme and are only meant to make the development process more
convenient.

+ Update `diff-mode' and `magit' to use the new faces.

This commit is the end-result of an initiative that was announced in
issue 18: https://gitlab.com/protesilaos/modus-themes/-/issues/18

The original idea was to test whether diffs could be improved by
removing most background elements.  Basically to just use red and green
for the affected lines, with red and green background used only for
refined diff views (in-line changes).

It eventually became clear that such a course of action would be to the
detriment of productivity, especially in Magit where interactivity (and
concomitant visual cues) are an essential aspect of the user experience.

So backgrounds are still in place.  And, in general, the approach is the
same as before.  The new colours are meant to be easier on the eyes
while always conforming with the WCAG AAA standard (minimum contrast
ratio of 7:1 between a given background and foreground combination).