Add support for CENTAUR-TABS
This addresses issue 25:
https://gitlab.com/protesilaos/modus-themes/-/issues/25

Note that I have experienced an issue with the colour of the bar that is
drawn on the current tab when setting `centaur-tabs-set-bar'.  The
colour does not change when switching from one theme to another (Modus
Operandi => Modus Vivendi, and vice-versa).  Whereas all other faces
adapt as expected.

My understanding is that the bar's colour is hard-coded, probably
inheriting from the active theme at the time the aforementioned variable
is first set.

I researched other themes that are linked for this package's README, to
see whether they provide a solution to this issue.  None of them seems
to have anything on the matter.  I thus assume there is some bug with
Centaur Tabs that I cannot troubleshoot as I am not a regular user.
Please check and report it accordingly.