REVERT 46c99ed: set background unconditionally
This addresses issue 27:
https://gitlab.com/protesilaos/modus-themes/-/issues/27

It reverts the commit that followed issue 26:
https://gitlab.com/protesilaos/modus-themes/-/issues/26

With regard to the latter, I recognise that the problem persists.  A
solution would need to be broader in scope.  I need expert help (or
documentation) to figure out the interplay between the terminal
emulator's or TTY's colours (first 16 ANSI sequences) and Emacs'
faces (the theme).  Any contribution to that end is more than welcome,
but changes from now on will be pushed to another branch.