Define new fringe colours and faces
These are meant to be applied to indicators that are displayed in the
fringes.

The motivation is to improve contrast with the colour of the fringe,
when the new customisation option is enabled by the user (see commit:
https://gitlab.com/protesilaos/modus-themes/-/commit/80fb704f83291acb23d646aa3e2463d78e5e3c93).

At any rate, these additions will allow us to improve things regardless
of that option.