Add customisation option for distinct Org blocks
This is off by default, as with all customisation options.

The idea is to render `org-mode' source blocks in a different background
than that of the main buffer.  Source block beginning and end lines are
configured to extend to the edge of the window when this option is set
to true.

Also tweak the background of the beginning and end lines of source
blocks, so that it differs from the block's contents.