Tweak `bg-paren-match' value and implementation
This concerns the colour that is applied as a background to the matching
delimiter with `show-paren-mode'.

Scenaria
--------

There are four scenaria of text-editing that need to be considered:

1. Paren match combined with the foreground of the delimiter.  The
contrast ratio for this should always be >= 7:1.

2. The foreground of the paren match needs to be 7:1 with the cursor
(while the cursor is the same as the main foreground).  This is because
Emacs will use the paren-match background as the character's foreground
when the cursor is placed over a highlighted delimiter.  Ergo, the
foreground value for scenario 1 needs to be the same as the main
foreground (pure black/white).

3. Paren match surrounded by the main background.  The minimum contrast
between two adjacent backgrounds can be as low as 1.25:1 (based on my
testing, which I am nevertheless willing to reconsider).  A slightly
higher value is preferable, while too high a value becomes annoying by
calling too much attention to itself.

4. Paren match within an active region.  Again the 1.25:1 applies.  Here
the ratio can approximate or meet the threshold because of the
interactive nature of marking a region.

Based on the above, we derive two constants to guide our process:

- fg-paren-match == fg-main
- bg-paren-match / bg-region == 1.25

The value for scenario 3 is the one left once the others are satisfied.

Contrast ratios
---------------

The table below shows the colour contrast of the old value of
`bg-paren-match' in the first row and the new values in the second one.
The idea is to have a more balanced result in all contexts where the
matching delimiter is highlighted.  Note that two themes do not have the
same contrast ratios, with the dark themes being lower, because the
human eye can spot more subtle differences on a dark background.

Modus Operandi:
|          |   #000000 |   #ffffff |  #bcbcbc |                       |
| bg-paren | (fg-main) | (bg-main) | (region) | Comment               |
|----------+-----------+-----------+----------+-----------------------|
| #deb8af  |      11.5 |      1.81 |     1.04 | Bad for region        |
| #efcabf  |      13.8 |      1.51 |     1.25 | Best for all contexts |

Modus Vivendi:
|          |   #ffffff |   #000000 |  #3c3c3c |                       |
| bg-paren | (bg-main) | (fg-main) | (region) | Comment               |
|----------+-----------+-----------+----------+-----------------------|
| #754a5d  |      7.25 |      2.89 |     1.52 | Fairly good           |
| #6e3a50  |      8.80 |      2.38 |     1.25 | Best for all contexts |

Finally, the bg-paren values are slightly accented towards a direction
that is not common for other foreground values that are applied to
syntax highlighting.  This is to keep things clear, whereas e.g. an
intense magenta could be conflated with the name of a keyword: not what
you want from some transient feedback.

This commit relates to the tweaks that followed the report in issue 29:
https://gitlab.com/protesilaos/modus-themes/-/issues/29