Add support for VC-ANNOTATE (but read commit msg)
It is impossible to support this command's background mode, as it would
violate the accessibility standard of this project (WCAG AAA).  While it
would be a shame to not support it at all given that it is built into
Emacs and is quite useful overall.

Quoting from the relevant note in the README:

	Due to the unique way `vc-annotate' (`C-x v g') applies colours,
    support for its background mode (`vc-annotate-background-mode') is
    disabled at the theme level.

	Normally, such a drastic measure should not belong in a theme:
    assuming the user's preferences is bad practice.  However, it has
    been deemed necessary in the interest of preserving colour contrast
    accessibility while still supporting a useful built-in tool.

	If there actually is a way to avoid such a course of action, without
    prejudice to the accessibility standard of this project, then please
    report as much (or contribute as per the information in the
    Contributing section).