Define new dedicated colours for language checkers
These shall apply to code linters as well as orthography and grammar
tools of various sorts.

As with all "dedicated colours", the intent is to not make any
compromises on the final outcome by being forced to reuse a colour whose
design was informed by other considerations (e.g. the blue for code
syntax highlighting has to work in that context, but may not be good
enough for another purpose).