Refine FLYMAKE: new linter colours, better fringes
Flymake does not define faces for its fringe indicators, offering
customisation options instead.  The problem with those is that they
inherit from faces that are not necessarily optimised for the fringes
and for the specific purpose of calling the user's attention to some
problem in the code.  As such, we set to customise those values at the
theme level, with the proviso that this should ideally be a standard set
of faces provided by Flymake.

For more on the new linter colours, see commit 53e1a43.