Add tentative support for Emacs27 TAB-{BAR,LINE}
The reason this is considered "tentative" is that I am using an uncommon
way to declare those faces: rather than add them to the list, I wrap
them in a new `custom-theme-set-faces' that evaluates for Emacs 27+.

Such a workaround is necessary to avoid errors with the linter.  It
throws an error every time a built-in package is introduced in a version
of Emacs that is higher than the minimum requirement specified in the
theme's header.