NEW CUSTOM OPTION: 3d modeline for active window
As with all customisation options, this is set to `nil' by default.

The intent is to apply a three-dimensional button effect to the focused
window's mode line.  Anyone who has ever used the standard Emacs theme
(emacs -Q) will immediately recognise the aesthetics.

The default is a flat rectangle with a border around it.  With this new
option set to `t', the overall impression of the focused window's
modeline is less pronounced.

For the design of this option, I benefited immensely from the guidance I
received from Omar Antolín Camarena <omar@matem.unam.mx> with regard to
the possibility of using a defun.  Any errors in execution, any
omissions or irregularities of any kind, are solely my own.