NEW CUSTOMISATION OPTION for "rainbow headings"
As with all customisation options, this is set to nil by default.

The idea with this toggle is to allow users to apply more vivid colours
to headings.  Currently, the effect is implemented in the Org and
Outline modes, though other face groups will be supported in the near
future.