NEW CUSTOMISATION OPTION for "section headings"
As with all customisation options, this is set to nil by default.

When enabled, it will make headings stand out more by applying to them a
subtle coloured background and an overline.  Faces that use this feature
right now are those of org-mode and outline-mode, though in principle
more can be added in the near future.

This option can be combined with all the existing ones that affect
headings.  Of notice is the recently-introduced "rainbow headings" (see
commit d3958bb) which will use a more saturated colour than the
default.