Always use a nuanced foreground for linted text
This is to ensure consistency between terminals that support wavy lines
and those that do not.

It also expands the relevant parts of commit a395411.  The issue with
Flycheck referenced therein is not for themes to solve.  Besides, the
use of "nuanced" colours for the underlined text should not pose an
impediment to editing.