Use less intense foreground for fringe indicators
This builds on the latest refinements to the overall luminance of the
colours used for fringe indicators and how these relate to relevant
faces.  See commits 4c18005 and 24cea55.

This changeset should further improve the looks of indicators such as
those produced by Flycheck in these cases:

https://github.com/flycheck/flycheck/pull/1743

My update is demonstrated here (also shows diff-hl and flymake):

https://gitlab.com/protesilaos/modus-themes/-/wikis/uploads/bb8e1098eb9c43878b111f6d2017cef4/Screenshot_from_2020-04-28_14-40-33.png

https://gitlab.com/protesilaos/modus-themes/-/wikis/uploads/6f91d74fbee80da6643a1bb0ae095b5f/Screenshot_from_2020-04-28_14-41-03.png