Make "section headings" customisation `:extend t'
This is for versions of Emacs that are >= 27.  The face's attributes,
such as the background colour, will now extend to the end of the window
rather than stop at the last character.  Such was the standard behaviour
in prior Emacs releases.

The idea for this change was recommended in issue 40:
https://gitlab.com/protesilaos/modus-themes/-/issues/40