Major review of colours for tabs. Tweak faces.
The palette intended for tabbed interfaces has been designed anew.  It
was essential to make these colours work in a broad range of scenaria,
such as by accounting for their style relative to a buffer's header
line or their foreground compared to that of a buffer.

The face groups for tab-bar and tab-line (Emacs 27) have been reviewed
to make better use of the new colours.  The centaur-tabs faces received
some minor tweaks as well (in the interest of consistency).

The present commit incorporates the feedback I received in issue 41:
https://gitlab.com/protesilaos/modus-themes/-/issues/41