Don't set background in regular ORG modeline clock
This fixes a mistake of mine.  The use of `bg-main' in the modeline
meant that the clock's area would look the same as that of the buffer.
Whereas now it will inherit the modeline's background (bg-active or
bg-inactive).

While the use of `fg-main' (pure black/white depending on the theme)
ensures that the clock has a text colour that is more intense than the
ordinary text on the modeline (though without having any accent that
would draw too much attention to it).