Major Review of fly{check,make,spell} for GUI/TUI
For GUI Emacs, all affected faces will now just use a colour-coded wavy
underline.  We no longer change the foreground value of the offending
text.  Whereas before the text would also get a colour, which was too
intrusive in most circumstances.

If support for wavy underlines is not available, we assume the presence
of a TUI, which generally is also more limited in its ability to
reproduce colours accurately.  So for those cases we apply a straight
underline combined with a colour-coded foreground for the affected
text.  This makes it more intense compared to the GUI equivalent, but is
the necessary course of action to overcome the constraints imposed by
the underlying terminal.