Review SELECTRUM faces (due to changes upstream)
Selectrum will place the *-highlight over selectrum-current-candidate,
so we can no longer use just a foreground value for the former as it
leads to inaccessible colour combinations.  A background+foreground
combination is needed instead.

To that end, the mapping of colours is also changed to better match the
aesthetics of this tool.