Define palette as a constant
I made this switch because I read '(info "(elisp) Defining Variables")',[0]
and was immediately drawn to the following reference:

    One notable difference between the two forms is that defconst
    unconditionally initializes the variable, whereas defvar initializes
    it only if it is originally void.

[0]:
https://www.gnu.org/software/emacs/manual/html_node/elisp/Defining-Variables.html#Defining-Variables

From my part, this fixes an issue that has persisted ever since commit
d6586a4f1, where a new colour value would not be read until after the
Emacs session was terminated.  With defconst, new values are read on
theme re-load.  This, in itself, is a win.  Because of that, I suspect
that the change may also fix issue 51:
https://gitlab.com/protesilaos/modus-themes/-/issues/51