No more need for ':extend t' in HELM faces
A recent commit in Helm adds the necessary ':extend' attribute to the
relevant faces (only relevant for >= Emacs27).  See
https://github.com/emacs-helm/helm/commit/d19592f593755a0d3daf09b4312daa0731d01dd6

This attribute is always retained regardless of the active theme, unless
it is explicitly overridden.  Meaning that support for it on the theme
side is no longer required.

Thanks to Manuel Uberti for bringing this issue to my attention:
https://github.com/emacs-helm/helm/issues/2213