Define missing 'header-line-highlight' face
This is meant to differentiate 'header-line-highlight' from 'highlight'.
The latter is ordinarily used in contexts where the background is not a
shade of grey and can, therefore, apply a more subtle shade of blue.
Whereas the 'header-line' has a more intense grey background (relative
to the main bg) and, therefore, requires a more pronounced blue.  The
same approach applies to the modeline.