Review 'org-meta-line', amending commit 8f8bb3b39
Use a more appropriate colour combination for this face.  The colours
used before gave a yellowish impression, whereas now we move towards the
blue side of the spectrum.  The latter is more consistent with the
various permutations of context where this face is used.

For context, the following piece of text is the message of the
aforementioned commit:

	This makes it look more like the lines that delimit Org blocks, but
    keeps them distinct enough.  The previous design would mimic
    comments in source code (faint foreground with an optional
    slant---when the "slanted constructs" options in enabled), which
    seemed less suitable for the special roles these constructs
    fulfil (e.g. holding a table formula or parsing the output of a code
    block).

	I am open to reviewing the new design if it proves problematic.