Major review of "nuanced" colours (internal)
The complete report is available on my website:
https://protesilaos.com/codelog/2020-07-08-modus-themes-nuanced-colours/

Quote from the link:

    The short version of this report is that the "nuanced" backgrounds
    have undergone a thoroughgoing review.  Now they are more consistent
    and better suited to their intended role, which is to provide a
    subtle coloured backdrop on which all existing foreground values can
    be used without prejudice to the overarching accessibility objective
    of the themes (namely, a contrast ratio that is >= 7:1, which
    corresponds to the highest standard of its kind: WCAG AAA).

This has implications for how the "section headings" option will style
the headings in Org buffers.  Some other faces are also affected, but
these may be less obvious on first sight.

This review opens up the possibility to properly address issue 34:
https://gitlab.com/protesilaos/modus-themes/-/issues/34