MAJOR REFACTOR of bold weight and bold constructs
The idea with all these changes is to provide the means for the user to
change the precise weight of a bold construct by modifying a single
face---that face is conveniently named 'bold'.  Whereas before the same
would have been laborious and prone to errors.

So assuming your font of choice has a semibold weight, you can now
change all instances of bold to actually be semibold.  Do it thus:

(set-face-attribute 'bold nil :weight 'semibold)

This commit should, in principle, address issue 63:
https://gitlab.com/protesilaos/modus-themes/-/issues/63