Add Info manual
The idea is to make it easier for users to read about the themes
directly from inside of Emacs.  If this works as expected, the README
will be de-prioritised and simply function as a placeholder that directs
users to this document instead.

At any rate, the HTML version is already updated to use the Info
manual's text: https://protesilaos.com/modus-themes