Initial implementation of reviewed headings option
The idea is to consolidate boolean options into a single alist.  It
shall allow users to control the styles of headings with greater
precision.

Given this opportunity, we also re-imagine the overall end-user
experience and expand the array of options, such as headings with normal
weight fonts, or just an overline.

Some complementary faces, such as Org's "TODO" keywords are reconsidered
as well.

THIS IS A WORK IN PROGRESS.  For an ongoing discussion see issue 81:
https://gitlab.com/protesilaos/modus-themes/-/issues/81.  Thanks to Adam
Spiers for the valuable feedback thus far!