Minor tweaks to custom modeline 'moody' option
* modus-operandi-theme.el (modus-operandi-theme-mode-line-attrs):
Add optional fg-distant.  Used as a distant-foreground property
when the 'moody' option is enabled.
(custom-theme-set-faces): Tweak mode-line and mode-line-inactive
faces.  The latter no longer uses a distant-foreground property,
in the interest of usability

* modus-vivendi-theme.el (modus-vivendi-theme-mode-line-attrs):
Add optional fg-distant.  Used as a distant-foreground property
when the 'moody' option is enabled.
(custom-theme-set-faces): Tweak mode-line and mode-line-inactive
faces.  The latter no longer uses a distant-foreground property,
in the interest of usability

For an ongoing discussion, refer to issue 80:
https://gitlab.com/protesilaos/modus-themes/-/issues/80.  Thanks to
Nicolas De Jaeghere for the valuable feedback!