Major review of Selectrum faces
* modus-operandi-theme.el, modus-vivendi-theme.el
(custom-theme-set-faces): Use bespoke styles for Selectrum's current
line to account for its design choice of always retaining the
colouration of matching characters.[0] We want to avoid scenaria where
matches are difficult to discern and the current line is not clear.

Review the remaining faces in the interest of consistency, but also to
align them with other pattern matching styles, like Orderless (which
Selectrum can use).

* modus-operandi-theme.el (modus-operandi-theme-completions)
(modus-operandi-theme-standard-completions)
(modus-operandi-theme-extra-completions): Tweak doc strings.

* modus-vivendi-theme.el (modus-vivendi-theme-completions)
(modus-vivendi-theme-standard-completions)
(modus-vivendi-theme-extra-completions): Tweak doc strings.

* doc/modus-themes.org (Option for completion framework aesthetics):
Clarify that, in terms of aesthetics, Selectrum is using styles that are
closer to Icomplete or Ido and, therefore, is not grouped together with
Helm and Ivy.

* doc/modus-themes.info: Recompile Info manual.

[0]: https://github.com/raxod502/selectrum/issues/21